user/requirements.txt
python-dotenv>=1.0.1
sqlmodel>=0.0.16
bcrypt>=4.1.2
fastapi>=0.110.1
sqlalchemy>=2.0.29
faker>=24.8.0
python-dateutil>=2.9.0.post0
pyjwt>=2.8.0
redis>=5.1.0b4